define puara-utils functions in header file
