+++ descriptors/puara-utils.h
@@ -49,9 +49,26 @@ namespace utils {
                 ) : current_value(currentValue), old_value(oldValue), leak(leakValue), 
                     frequency(freq), timer(timerValue) {}
 
-            double integrate(double reading);
-            double integrate(double reading, double leak);
-            double integrate(double reading, double leak, long long time);
+            double integrate(double reading) {
+                double leakValue = custom_leak;
+                if (frequency <= 0) {
+                    current_value = reading + (old_value * leakValue);
+                } else if ((time/1000LL)  - (1000 / frequency) < timer) {  
+                    current_value = reading + old_value;
+                } else {
+                    current_value = reading + (old_value * leakValue);
+                    timer = (time/1000LL);
+                }
+                return current_value;
+            }
+
+            double integrate(double reading, double leak) {
+                return LeakyIntegrator::integrate (reading, leak, getCurrentTimeMicroseconds());
+            }
+
+            double integrate(double reading, double leak, long long time) {
+                return LeakyIntegrator::integrate (reading, leak);
+            }
     };
 
     /**
@@ -65,9 +82,24 @@ namespace utils {
             double inMax = 0;
             double outMin = 0;
             double outMax = 0;
-            double range (double in);
-            float range (float in);
-            int range (int in);
+            double range (double in) {
+                current_in = in;
+                if (outMin != outMax) {
+                    return (in - inMin) * (outMax - outMin) / (inMax - inMin) + outMin;
+                } else {
+                    return in;
+                }
+            }
+
+            float range (float in) {
+                double casted_in = static_cast<double>(in);
+                return static_cast<float>(range(casted_in));
+            }
+
+            int range (int in) {
+                double casted_in = static_cast<double>(in);
+                return static_cast<int>(range(casted_in));
+            }
     };
 
     /**
@@ -78,7 +110,13 @@ namespace utils {
         public:
             int size = 10;
             std::deque<double> buffer;
-            double add(double element);
+            double add(double element) {
+                buffer.push_front(element);
+                if (buffer.size() > size) {
+                    buffer.pop_back();
+                }
+                return element;
+            }
     };
 
     template <typename T>
@@ -92,7 +130,17 @@ namespace utils {
             RollingMinMax(size_t buffer_size=10) : buf(buffer_size) {}
 
             puara_gestures::MinMax<T> current_value;
-            puara_gestures::MinMax<T> update(T value);
+            puara_gestures::MinMax<T> update(T value) {
+                puara_gestures::MinMax<T> ret{.min = value, .max = value};
+                buf.push_back(value);
+                for(const T value : buf) {
+                    if(value < ret.min) ret.min = value;
+                    if(value > ret.max) ret.max = value;
+                }
+                current_value = ret;
+                return ret;
+            }
+
         private:
             boost::circular_buffer<T> buf;
     };
@@ -100,19 +148,51 @@ namespace utils {
     /**
      *  @brief Simple function to get the current elapsed time in microseconds.
      */ 
-    long long getCurrentTimeMicroseconds();
+    long long getCurrentTimeMicroseconds() {
+        auto currentTimePoint = std::chrono::high_resolution_clock::now();
+        auto duration = std::chrono::duration_cast<std::chrono::microseconds>(currentTimePoint.time_since_epoch());
+        return duration.count();
+    }
 
     /**
      * @brief Function used to reduce feature arrays into single values. 
      * E.g., brush uses it to reduce multiBrush instances
      */
-    double arrayAverageZero (double * Array, int ArraySize);
+    double arrayAverageZero (double * Array, int ArraySize) {
+        double sum = 0;
+        int count = 0;
+        double output = 0;
+        for (int i = 0; i < ArraySize; ++i) {
+            if (Array[i] != 0) {
+                sum += Array[i];
+                count++;
+            }
+        }
+        if (count > 0) {
+            output = sum / count; 
+        }
+        return output;
+    }
 
     /**
      * @brief Legacy function used to calculate 1D blob detection in older 
      * digital musical instruments
      */ 
-    void bitShiftArrayL (int * origArray, int * shiftedArray, int arraySize, int shift);
+    void bitShiftArrayL (int * origArray, int * shiftedArray, int arraySize, int shift) {
+        for (int i=0; i < arraySize; ++i) {
+            shiftedArray[i] = origArray[i];
+        }
+        for (int k=0; k < shift; ++k) {
+            for (int i=0; i < arraySize; ++i) {
+                if ( i == (arraySize-1)) {
+                    shiftedArray[i] = (shiftedArray[i] << 1);
+                }
+                else {
+                    shiftedArray[i] = (shiftedArray[i] << 1) | (shiftedArray[i+1] >> 7);
+                }
+            }
+        }
+    }
 }
 
 namespace convert {
@@ -121,37 +201,49 @@ namespace convert {
      * @brief Convert g's to m/s^2
      * 
      */
-    double g_to_ms2(double reading);
+    double g_to_ms2(double reading) {
+        return reading * 9.80665;
+    }
 
     /**
      * @brief Convert m/s^2 to g's
      * 
      */
-    double ms2_to_g(double reading);
+    double ms2_to_g(double reading) {
+        return reading / 9.80665;
+    }
 
     /**
      * @brief Convert DPS to radians per second
      * 
      */
-    double dps_to_rads(double reading);
+    double dps_to_rads(double reading) {
+        return reading * M_PI / 180;
+    }
 
     /**
      * @brief Convert radians per second to DPS
      * 
      */
-    double rads_to_dps(double reading);
+    double rads_to_dps(double reading) {
+        return reading * 180 / M_PI;
+    }
 
     /**
      * @brief Convert Gauss to uTesla
      * 
      */
-    double gauss_to_utesla(double reading);
+    double gauss_to_utesla(double reading) {
+        return reading / 10000;
+    }
 
     /**
      * @brief Convert uTesla to Gauss 
      * 
      */
-    double utesla_to_gauss(double reading);
+    double utesla_to_gauss(double reading) {
+        return reading * 10000;
+    }
 
 }
 }
